requirements.txt
rosbags
pyyaml
pandas
numpy
meshcat
meshcat-shapes